feat(kafka): добавление сервиса для отправки пользовательских событий и улучшение конфигурации
- Создан новый сервис `UserEventsService` для отправки пользовательских событий в Kafka.
