+++ Java/seminar8/src/main/java/backend/academy/kafka/config/CommonKafkaConfig.java
@@ -26,7 +26,7 @@ public class CommonKafkaConfig {
     private final KafkaProperties kafkaProperties;
 
     @Bean
-    Admin localKafkaClusterAdminClint() {
+    Admin localKafkaClusterAdminClient() {
         return AdminClient.create(
             Map.of(
                 AdminClientConfig.BOOTSTRAP_SERVERS_CONFIG,
