+++ Java/seminar8/src/main/java/backend/academy/kafka/controller/UserEventsController.java
@@ -1,15 +1,8 @@
 package backend.academy.kafka.controller;
 
-import static backend.academy.kafka.model.UserEvent.UserEventType.ACCRUAL;
-import backend.academy.kafka.model.UserEvent;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import java.time.LocalDateTime;
-import java.util.concurrent.atomic.AtomicLong;
+import backend.academy.kafka.service.UserEventsService;
 import lombok.RequiredArgsConstructor;
-import lombok.SneakyThrows;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.kafka.core.KafkaTemplate;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
@@ -21,12 +14,7 @@ import org.springframework.web.bind.annotation.RestController;
 @RequiredArgsConstructor
 public class UserEventsController {
 
-    private final AtomicLong idGenerator = new AtomicLong();
-    private final KafkaTemplate<Long, String> template;
-    private final ObjectMapper objectMapper;
-
-    @Value("${app.user-events.topic}")
-    private String topic;
+    private final UserEventsService service;
 
     @PostMapping("/users/{user-id}/events")
     public void sendMessages(
@@ -35,33 +23,7 @@ public class UserEventsController {
         @RequestParam(name = "force", defaultValue = "false") boolean force,
         @RequestParam(name = "transactional", defaultValue = "false") boolean transactional
     ) {
-        log.info("Отправляем {} сообщений по клиенту с ИД {}", count, userId);
-
-        if (transactional && template.isTransactional()) {
-            template.executeInTransaction(ops -> {
-                for (int i = 0; i < count; i++) {
-                    ops.send(topic, userId, createRandomEvent(userId));
-                }
-                return true;
-            });
-        } else {
-            for (int i = 0; i < count; i++) {
-                template.send(topic, userId, createRandomEvent(userId));
-            }
-            if (force) {
-                template.flush();
-            }
-        }
-    }
-
-    @SneakyThrows
-    private String createRandomEvent(long userId) {
-        return objectMapper.writeValueAsString(
-            new UserEvent()
-                .setUserId(userId)
-                .setType(ACCRUAL)
-                .setCreatedAt(LocalDateTime.now())
-                .setId(idGenerator.incrementAndGet()));
+        service.sendMessages(userId, count, force, transactional);
     }
 
 }
