+++ Java/seminar8/src/main/resources/README.md
@@ -3,3 +3,4 @@
 #### Поелзные ссылки
 - [Swagger](http://localhost:8080/swagger-ui/index.html#/user-events-controller/sendMessages)
 - [Kafka UI](http://localhost:8082/)
+- Директория с данными Kafka - `/var/lib/kafka/data`
